refactored free route
